requirements.txt
djangorestframework==3.12.4
django-filter==2.4.0
Django>=3.2.9,<4.0
psycopg2-binary>=2.9
django-cors-headers
psycopg2